Include root items with property class names
Before this change, `Item.node_set_from` would accidentally ignore root
elements that _also_ had a property class name.

```html
<body>
  <div class="h-card p-author">
    …
  </div>
</body>
```

The original implementation would see `p-author` as a property class
name and _not_ push the `context` onto the `node_set`. This change
resolves that by looking at the node's ancestors to see whether or not
it has a "parent" item.

So, if the node is an item node _and_ it has a property class name, but
it _does not_ have an ancestor that is also an item node, add it to
`node_set`.

Phew.